requirements.txt
fastapi==0.122.0
llama-index-llms-groq==0.4.1
numpy==2.3.5
pandas==2.3.3
python-dotenv==1.2.1
tqdm==4.67.1
uvicorn==0.38.0
torch==2.9.1

